Started to move the messages storage into liblinphone

+++ src/org/linphone/ChatStorage.java
@@ -22,6 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.linphone.core.LinphoneChatMessage;
+import org.linphone.core.LinphoneChatRoom;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -39,22 +40,41 @@ public class ChatStorage {
 	private static final int OUTGOING = 0;
 	private static final int READ = 1;
 	private static final int NOT_READ = 0;
+	
+	private static ChatStorage instance;
 	private Context context;
 	private SQLiteDatabase db;
+	private boolean useNativeAPI;
 	private static final String TABLE_NAME = "chat";
 	private static final String DRAFT_TABLE_NAME = "chat_draft";
 
-	public ChatStorage(Context c) {
+	public synchronized static final ChatStorage getInstance() {
+		if (instance == null)
+			instance = new ChatStorage(LinphoneService.instance().getApplicationContext());
+		return instance;
+	}
+	
+	private ChatStorage(Context c) {
 	    context = c;
-	    ChatHelper chatHelper = new ChatHelper(context);
-	    db = chatHelper.getWritableDatabase();
+	    useNativeAPI = c.getResources().getBoolean(R.bool.use_linphone_chat_storage);
+	    
+	    if (!useNativeAPI) {
+		    ChatHelper chatHelper = new ChatHelper(context);
+		    db = chatHelper.getWritableDatabase();
+	    }
 	}
 	
 	public void close() {
-		db.close();
+		if (!useNativeAPI) {
+			db.close();
+		}
 	}
 	
 	public void updateMessageStatus(String to, String message, int status) {
+		if (useNativeAPI) {
+			return;
+		}
+		
 		String[] whereArgs = { String.valueOf(OUTGOING), to, message };
 		Cursor c = db.query(TABLE_NAME, null, "direction LIKE ? AND remoteContact LIKE ? AND message LIKE ?", whereArgs, null, null, "id DESC");
 
@@ -75,6 +95,10 @@ public class ChatStorage {
 	}
 	
 	public void updateMessageStatus(String to, int id, int status) {
+		if (useNativeAPI) {
+			return;
+		}
+		
 		ContentValues values = new ContentValues();
 		values.put("status", status);
 		
@@ -82,6 +106,10 @@ public class ChatStorage {
 	}
 	
 	public int saveTextMessage(String from, String to, String message, long time) {
+		if (useNativeAPI) {
+			return -1;
+		}
+		
 		ContentValues values = new ContentValues();
 		if (from.equals("")) {
 			values.put("localContact", from);
@@ -102,6 +130,10 @@ public class ChatStorage {
 	}
 	
 	public int saveImageMessage(String from, String to, Bitmap image, String url, long time) {
+		if (useNativeAPI) {
+			return -1;
+		}
+		
 		ContentValues values = new ContentValues();
 		if (from.equals("")) {
 			values.put("localContact", from);
@@ -129,6 +161,11 @@ public class ChatStorage {
 	}
 	
 	public void saveImage(int id, Bitmap image) {
+		if (useNativeAPI) {
+			//TODO
+			return;
+		}
+		
 		if (image == null)
 			return;
 		
@@ -141,6 +178,11 @@ public class ChatStorage {
 	}
 	
 	public int saveDraft(String to, String message) {
+		if (useNativeAPI) {
+			//TODO
+			return -1;
+		}
+		
 		ContentValues values = new ContentValues();
 		values.put("remoteContact", to);
 		values.put("message", message);
@@ -148,6 +190,11 @@ public class ChatStorage {
 	}
 	
 	public void updateDraft(String to, String message) {
+		if (useNativeAPI) {
+			//TODO
+			return;
+		}
+		
 		ContentValues values = new ContentValues();
 		values.put("message", message);
 		
@@ -155,10 +202,20 @@ public class ChatStorage {
 	}
 	
 	public void deleteDraft(String to) {
+		if (useNativeAPI) {
+			//TODO
+			return;
+		}
+		
 		db.delete(DRAFT_TABLE_NAME, "remoteContact LIKE \"" + to + "\"", null);
 	}
 	
 	public String getDraft(String to) {
+		if (useNativeAPI) {
+			//TODO
+			return "";
+		}
+		
 		Cursor c = db.query(DRAFT_TABLE_NAME, null, "remoteContact LIKE \"" + to + "\"", null, null, null, "id ASC");
 
 		String message = null;
@@ -176,17 +233,22 @@ public class ChatStorage {
 	
 	public List<String> getDrafts() {
 		List<String> drafts = new ArrayList<String>();
-		Cursor c = db.query(DRAFT_TABLE_NAME, null, null, null, null, null, "id ASC");
+		
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			Cursor c = db.query(DRAFT_TABLE_NAME, null, null, null, null, null, "id ASC");
 
-		while (c.moveToNext()) {
-			try {
-				String to = c.getString(c.getColumnIndex("remoteContact"));
-				drafts.add(to);
-			} catch (Exception e) {
-				e.printStackTrace();
+			while (c.moveToNext()) {
+				try {
+					String to = c.getString(c.getColumnIndex("remoteContact"));
+					drafts.add(to);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
+			c.close();
 		}
-		c.close();
 		
 		return drafts;
 	}
@@ -194,94 +256,140 @@ public class ChatStorage {
 	public List<ChatMessage> getMessages(String correspondent) {
 		List<ChatMessage> chatMessages = new ArrayList<ChatMessage>();
 		
-		Cursor c = db.query(TABLE_NAME, null, "remoteContact LIKE \"" + correspondent + "\"", null, null, null, "id ASC");
-		
-		while (c.moveToNext()) {
-			try {
-				String message, timestamp, url;
-				int id = c.getInt(c.getColumnIndex("id"));
-				int direction = c.getInt(c.getColumnIndex("direction"));
-				message = c.getString(c.getColumnIndex("message"));
-				timestamp = c.getString(c.getColumnIndex("time"));
-				int status = c.getInt(c.getColumnIndex("status"));
-				byte[] rawImage = c.getBlob(c.getColumnIndex("image"));
-				int read = c.getInt(c.getColumnIndex("read"));
-				url = c.getString(c.getColumnIndex("url"));
-				
-				ChatMessage chatMessage = new ChatMessage(id, message, rawImage, timestamp, direction == INCOMING, status, read == READ);
-				chatMessage.setUrl(url);
+		if (!useNativeAPI) {
+			Cursor c = db.query(TABLE_NAME, null, "remoteContact LIKE \"" + correspondent + "\"", null, null, null, "id ASC");
+			
+			while (c.moveToNext()) {
+				try {
+					String message, timestamp, url;
+					int id = c.getInt(c.getColumnIndex("id"));
+					int direction = c.getInt(c.getColumnIndex("direction"));
+					message = c.getString(c.getColumnIndex("message"));
+					timestamp = c.getString(c.getColumnIndex("time"));
+					int status = c.getInt(c.getColumnIndex("status"));
+					byte[] rawImage = c.getBlob(c.getColumnIndex("image"));
+					int read = c.getInt(c.getColumnIndex("read"));
+					url = c.getString(c.getColumnIndex("url"));
+					
+					ChatMessage chatMessage = new ChatMessage(id, message, rawImage, timestamp, direction == INCOMING, status, read == READ);
+					chatMessage.setUrl(url);
+					chatMessages.add(chatMessage);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+			c.close();
+		} else {
+			LinphoneChatRoom room = LinphoneManager.getLc().createChatRoom(correspondent);
+			for (LinphoneChatMessage message : room.getHistory()) {
+				ChatMessage chatMessage = new ChatMessage(message.hashCode(), message.getText(), null, String.valueOf(message.getTime()), true, 0, true);
+				chatMessage.setUrl(message.getExternalBodyUrl());
 				chatMessages.add(chatMessage);
-			} catch (Exception e) {
-				e.printStackTrace();
 			}
 		}
-		c.close();
 		
 		return chatMessages;
 	}
 
 	public String getTextMessageForId(int id) {
-		Cursor c = db.query(TABLE_NAME, null, "id LIKE " + id, null, null, null, null);
-
 		String message = null;
-		if (c.moveToFirst()) {
-			try {
-				message = c.getString(c.getColumnIndex("message"));
-			} catch (Exception e) {
-				e.printStackTrace();
+		
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			Cursor c = db.query(TABLE_NAME, null, "id LIKE " + id, null, null, null, null);
+	
+			if (c.moveToFirst()) {
+				try {
+					message = c.getString(c.getColumnIndex("message"));
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
+			c.close();
 		}
-		c.close();
 		
 		return message;
 	}
 	
 	public void removeDiscussion(String correspondent) {
-		db.delete(TABLE_NAME, "remoteContact LIKE \"" + correspondent + "\"", null);
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			db.delete(TABLE_NAME, "remoteContact LIKE \"" + correspondent + "\"", null);
+		}
 	}
 	
 	public ArrayList<String> getChatList() {
 		ArrayList<String> chatList = new ArrayList<String>();
 		
-		Cursor c = db.query(TABLE_NAME, null, null, null, "remoteContact", null, "id DESC");
-		while (c != null && c.moveToNext()) {
-			try {
-				String remoteContact = c.getString(c.getColumnIndex("remoteContact"));
-				chatList.add(remoteContact);
-			} catch (IllegalStateException ise) {
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			Cursor c = db.query(TABLE_NAME, null, null, null, "remoteContact", null, "id DESC");
+			while (c != null && c.moveToNext()) {
+				try {
+					String remoteContact = c.getString(c.getColumnIndex("remoteContact"));
+					chatList.add(remoteContact);
+				} catch (IllegalStateException ise) {
+				}
 			}
+			c.close();
 		}
-		c.close();
 		
 		return chatList;
 	}
 
 	public void deleteMessage(int id) {
-		db.delete(TABLE_NAME, "id LIKE " + id, null);
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			db.delete(TABLE_NAME, "id LIKE " + id, null);
+		}
 	}
 	
 	public void markMessageAsRead(int id) {
-		ContentValues values = new ContentValues();
-		values.put("read", READ);
-		db.update(TABLE_NAME, values, "id LIKE " + id, null);
+		if (useNativeAPI) {
+			//TODO
+		} else {
+			ContentValues values = new ContentValues();
+			values.put("read", READ);
+			db.update(TABLE_NAME, values, "id LIKE " + id, null);
+		}
 	}
 	
 	public int getUnreadMessageCount() {
-		Cursor c = db.query(TABLE_NAME, null, "read LIKE " + NOT_READ, null, null, null, null);
-		int count = c.getCount();
-		c.close();
+		int count;
+		if (!useNativeAPI) {
+			Cursor c = db.query(TABLE_NAME, null, "read LIKE " + NOT_READ, null, null, null, null);
+			count = c.getCount();
+			c.close();
+		} else {
+			//TODO
+			count = -1;
+		}
 		return count;
-		
 	}
 
 	public int getUnreadMessageCount(String contact) {
-		Cursor c = db.query(TABLE_NAME, null, "remoteContact LIKE \"" + contact + "\" AND read LIKE " + NOT_READ, null, null, null, null);
-		int count = c.getCount();
-		c.close();
+		int count;
+		if (!useNativeAPI) {
+			Cursor c = db.query(TABLE_NAME, null, "remoteContact LIKE \"" + contact + "\" AND read LIKE " + NOT_READ, null, null, null, null);
+			count = c.getCount();
+			c.close();
+		} else {
+			//TODO
+			count = -1;
+		}
 		return count;
 	}
 
 	public byte[] getRawImageFromMessage(int id) {
+		if (useNativeAPI) {
+			//TODO
+			return null;
+		}
+		
 		String[] columns = { "image" };
 		Cursor c = db.query(TABLE_NAME, columns, "id LIKE " + id + "", null, null, null, null);
